Hi,
here's a patch for Net::LDAP. I tried to connect to an LDAP
server, but it failed within IO::Socket::SSL with a "Invalid
certificate authority locations" error. However,
IO::Socket::SSL only stores those kinds of errors in
$IO::Socket::SSL::SSL_ERROR, which Net::LDAP did not check,
so all I got was an "Operations error". This patch should
fix that.