[PATCH] use of non-localised $_ in Net::LDAP::Constant causes prolems in lemonldap-ng
[perl-ldap 0.4001; perl 5.10.1]

Please find attached a patch that fixes a problem when requiring
Net::LDAP in certain environments. The problem can be demonstrated by
running:

perl -we'require Net::LDAP for (1,2)'

Similar code is used in lemonldap-ng's test suite since a long time,
and with perl-ldap 0.4001, tests suddenly fail. There is a bug report
in Debian about this at http://bugs.debian.org/577340

I hope the patch is OK and you would incorporate it in a future
perl-ldap release.