LDAP.{pm,pod} new method done() as alias for unbind()
This alias helps to clean up the object the same way no matter whether
it is a Net::LDAP or a Net::LDAP::LDIF object.

This is mainly intended as a convenience function in cooperation
with Net::LDAP::Entry->update().

After the creation of the Net::LDAP resp. Net::LDAP::LDIF objects using
their respective constructors, changes can be done consistently using
Net::LDAP::Entry->update() and the objects can be cleaned up using their
bind() method.

This comes extremely handy when writing a client that offers the option
to simulate writes to the directory.
Simply use the Net::LDAP::LDIF object with the methods above the same way
you would use the Net::LDAP object, and you will get an LDIF file with
the simulated changes.