RT#81380: in Filter.pm unescape hexpairs and ( ) \ * only
Fix long-standing bug that affects filter values containing backslashes.
Instead of unescaping all characters prefixed by a backslash, restrict
the unescaping to the characters with special meaning in filters.

While not fully conforming to the RFC [which only allows unescaping hexpairs],
this is now a lot better than the original implementation.