Present the name provided by the caller for GSSAPI/SASL authentication
Reverts a change from af630673 which attempted to work around the use of
round-robin A records by always passing the IP address of the LDAP server to
SASL/GSSAPI which caused a reliance on Kerberos DNS canonicalization to guess
the correct service name for authentication. This causes problems with Kerberos
configurations that do not perform DNS canonicalization -- the service name
cannot be determined despite the fact that it was correctly provided by the
caller.