Present the name provided by the caller for GSSAPI/SASL authentication
Reverts a change from af63067 which attempted to work around the use of
round-robin A records by always passing the IP address of the LDAP server to
SASL/GSSAPI which caused a reliance on Kerberos DNS canonicalization to guess
the correct service name for authentication. This causes problems with Kerberos
configurations that do not perform DNS canonicalization -- the service name
cannot be determined despite the fact that it was correctly provided by the
caller.

Provides for a new option which can be passed to bind which will force us to
continue to use the old behavior, but by default, uses the hostname as provided
by the caller for authentication purposes.

To reproduce the old behavior, instead of:

    $ldap->bind(...)

Use:

    $ldap->bind(... sasl_authenticate_with => 'ip')