RT#91210: Paged.pm: fix example code
RFC 2696 explicitly states that only a cookie set to the empty string
indicates that the search is done. I.e. a cookie value of '0' is allowed
and must not terminate the search.

Adapt the example code to reflect this fact.

Thanks to David De Borre for reporting the bug and the 389 Directory Server
team for the analysis.