Remove Test Requirement to Text::Soundex
The test requirement to Text::Soundex causes the need to install a module that needs a compiler to be built. As the general project guidelines seem to target a mostly perl only based install, this should be considered for removal. As far as I can see, Text::Soundex is only an option to the Net::LDAP::FilterMatch, so there is no need to require it here.